refactor escapers
- fold UnicodeEscaper into PercentEscaper
- remove Java dependencies in PercentEscaper
- make both classes `internal`
